--- Project/test_main_smart_meter.py
@@ -1,20 +1,0 @@
-from datetime import datetime, timedelta
-import time
-import unittest
-import sys
-
-from main_smart_meter import meter
-
-class TestStringMethods(unittest.TestCase):
-
-                                                                                                                        
-    mainmodule = meter()
-
-    def test_dummy(self):
-        self.assertEqual(True, True)
-
-
-
-if __name__ == '__main__':
-    unittest.main()
-
